Rename a package.
